move drag scrolling handler to WaveformFrame
